+++ Sources/VComponents/Components/Buttons/Rounded Button/VRoundedButtonUIModel.swift
@@ -69,14 +69,14 @@ public struct VRoundedButtonUIModel {
         /// Icon size.
         /// Set to `20x20` on `iOS`.
         /// Set to `14x14` on `macOS`.
-        /// Set to `20x20` on `watchOS`.
+        /// Set to `24x24` `watchOS`.
         public var iconSize: CGSize = {
 #if os(iOS)
             return CGSize(dimension: 20)
 #elseif os(macOS)
             return CGSize(dimension: 14)
 #elseif os(watchOS)
-            return CGSize(dimension: 20)
+            return CGSize(dimension: 24)
 #else
             fatalError() // Not supported
 #endif