+++ Sources/VComponents/Components/Buttons/Stretched Button/VStretchedButtonState.swift
@@ -1,0 +1,14 @@
+//
+//  VStretchedButtonState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 03.04.23.
+//
+
+import Foundation
+import VCore
+
+// MARK: - V Stretched Button Internal State
+/// Enum that represents state, such as `enabled`, `pressed`, or `disabled`.
+@available(tvOS, unavailable)
+public typealias VStretchedButtonInternalState = GenericState_EnabledPressedDisabled